Added Player Blink/Teleport ability
The player can now blink/teleport to a preset blinkPoint GameObject.  This ability can be activated by pressing either the 'Q' key or by pressing the right bumper button on any game pad.

The player can not blink into other collider but can blink through some colliders (like walls).  This ability could result in some glitches where the player can blink/teleport out of bounds.

Even though the code for this ability is in place I would recommend that we focus on getting the other parts of the character controller up and running before focusing on this ability.