Added Health/Damage System
The player now has a set health with a minimum and maximum.  This heath can be changed either via code or by having the player collide into a collider with a damage script attached.  This script has a single variable on it called units.  The number set there will be added to the players current health upon impact.  I.E. 50 units will increase the players health by 50 (this could be used for healing items), or -50 units will decrease the players health by 50 units (This could be used for enemy's) .   The collider that the script is attached to can be a standard collider or a trigger, either one will work.