Added Health Packs
The player can now hold a set number of health packs and use them to replenish his health.  The health packs can be used with the 'e' key or the left bumper button on any gamepad.  I also added a health pack prefab example.  This will destroy itself when the player touches it and increase the number of health packs the player holds by 1.  It will not add more than the maximum amount you can carry.