Updated Player Death and Respawn system
TLDR: Added checkpoints and Updated Death system

Old Version:  Previously when the player died the Player game object destroyed itself and needed to be re-instantiated.  

New Version:  Now when the player dies the Player game object is not destroyed but is instead moved back to the last set check point.

Checkpoints:  A new checkpoint prefab has been added to the project.  This prefab looks like a transparent blue square in the scene view.  This blue square shows the trigger zone of said checkpoint and can be resized as needed.  The blue square sprite will destroy itself once the game enters play-mode (leaving the rest of the checkpoint intact and invisible) and is only for ease of development.  

Other notes: destroyOnImpact script has been renamed to DestroyObject.  The player controller has new events related to the new Death and Respawn system.